+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Interop/AzureMapsInterop.cs
@@ -129,6 +129,13 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop
         /// </summary>
         /// <returns></returns>
         ValueTask DisposeAsync();
+
+        /// <summary>
+        /// Removes a map instance.
+        /// </summary>
+        /// <param name="mapId"></param>
+        /// <returns></returns>
+        ValueTask RemoveMap(string mapId);
     }
 
     internal class AzureMapsInterop : IAzureMapsInterop, IAsyncDisposable
@@ -215,6 +222,14 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop
             var mapRef = await module.InvokeAsync<IJSObjectReference>(GetJsInteropMethod(), mapId.Trim());
 
             return mapRef;
+        }
+
+        public async ValueTask RemoveMap(string mapId)
+        {
+            ArgumentNullException.ThrowIfNullOrWhiteSpace(mapId, nameof(mapId));
+
+            var module = await moduleTask.Value;
+            await module.InvokeVoidAsync(GetJsInteropMethod(), mapId.Trim());
         }
 
         public async ValueTask SetLogLevel(LogLevel logLevel)