+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/scripts/modules/Atlas.ts
@@ -4,11 +4,11 @@ import { Logger, LogLevel } from "./common/"
 export class Atlas {
     public static setLanguage(language: string) {
         atlas.setLanguage(language);
-        Logger.logMessage("Setting Language", LogLevel.Information, `Language set to [${language}].`);
+        Logger.logMessage("Setting Language", LogLevel.Debug, `Language set to [${language}].`);
     }
 
     public static setView(view: string) {
         atlas.setView(view);
-        Logger.logMessage("Setting View", LogLevel.Information, `View set to [${view}].`);
+        Logger.logMessage("Setting View", LogLevel.Debug, `View set to [${view}].`);
     }
 }